pyproject.toml
[tool.poetry]
name = "data-stack"
version = "0.1.0"
description = ""
authors = ["claudiohpontes <claudiohpontes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
isort = "^5.13.2"
black = "^24.4.2"
taskipy = "^1.12.2"
mkdocs = "^1.6.0"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.5.19"
mkdocstrings = {extras = ["python"], version = "^0.25.0"}
requests = "^2.31.0"
tweepy = "^4.14.0"
pandas = "^2.2.2"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"